MainContainer extracted to new separate class
